+++ elliottlib/constants.py
@@ -3,7 +3,6 @@ This file contains constants that are used to manage OCP Image and RPM builds
 """
 
 BREW_HUB = "https://brewhub.engineering.redhat.com/brewhub"
-BREW_IMAGE_HOST = "brew-pulp-docker01.web.prod.ext.phx2.redhat.com:8888"
 CGIT_URL = "http://pkgs.devel.redhat.com/cgit"
 
 VALID_BUG_STATES = ['NEW', 'ASSIGNED', 'POST', 'MODIFIED', 'ON_QA', 'VERIFIED', 'RELEASE_PENDING', 'CLOSED']
